Moved the communicator to a library project and create an Interface for the communicator.
