+++ Shared/Communicator.cs
@@ -4,22 +4,22 @@ using System.Diagnostics;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.SignalR.Client;
 
-namespace WPFClient
+namespace Shared
 {
-    class Communicator
+    internal class Communicator :ICommunicator
     {
         private readonly HubConnection connection;
         private readonly Dictionary<string, Func<string, Task>> callBackTopics;
 
-        public Communicator( )
+        public Communicator()
         {
             callBackTopics = new Dictionary<string, Func<string, Task>>();
-            
+
             connection = new HubConnectionBuilder()
                 .WithUrl("https://localhost:44392/testhub")
                 .Build();
 
-            _ = Task.Run(async() => { await connection.StartAsync(); });
+            _ = Task.Run(async () => { await connection.StartAsync(); });
             connection.Closed += async (error) =>
             {
                 await Task.Delay(new Random().Next(0, 5) * 1000);
@@ -34,16 +34,16 @@ namespace WPFClient
         private void OnSubscribe(string response)
         {
             Debug.WriteLine(response);
-         }
+        }
 
         private void OnUnsubscribe(string response)
         {
             Debug.WriteLine(response);
         }
 
-        private void OnPublish(string topic, string content)
+        private async Task OnPublish(string topic, string content)
         {
-            callBackTopics[topic](content);
+            await callBackTopics[topic](content);
         }
         public async Task<string> SubscribeAsync(string topic, Func<string, Task> callBack)
         {
@@ -58,7 +58,7 @@ namespace WPFClient
 
         public async Task<string> UnsubscribeAsync(string topic)
         {
-            if(!callBackTopics.ContainsKey(topic))
+            if (!callBackTopics.ContainsKey(topic))
                 return "fail";
 
             await connection.InvokeAsync("UnsubscribeTopicAsync", topic);
@@ -66,7 +66,7 @@ namespace WPFClient
             return "success";
         }
 
-        public async Task PublishAsync(String topic, String message )
+        public async Task PublishAsync(String topic, String message)
         {
             await connection.InvokeAsync("PublishAsync", topic, message);
         }
